ui_application/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
confluent-kafka==2.2.0
dataclasses-json==0.5.14
emoji==2.8.0
environs==9.5.0
fastapi==0.103.1
filelock==3.12.4
filetype==1.2.0
frozenlist==1.4.0
fsspec==2023.9.1
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.2
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.285
langsmith==0.0.40
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
milvuskafka @ file:///Users/jbyrne/development/zilliz_kafka
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.3
pydantic==2.3.0
pydantic_core==2.6.3
pymilvus==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==5.3.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
unstructured==0.10.12
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
